fix(webchat): optional subtitle
